Fix some include paths after rename/refactor
