+++ Gems/Multiplayer/Code/Source/Debug/MultiplayerDebugSystemComponent.cpp
@@ -10,7 +10,7 @@
 *
 */
 
-#include <Source/Imgui/MultiplayerDebugSystemComponent.h>
+#include <Source/Debug/MultiplayerDebugSystemComponent.h>
 #include <AzCore/Serialization/SerializeContext.h>
 #include <AzCore/Interface/Interface.h>
 #include <Include/IMultiplayer.h>
